Extract checkstyle output into a separate module.
Rename functions a bit now that they are specific to checkstyle output.
